move stringToJsRegex
